ubicua/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
paho-mqtt==1.6.1
psycopg-binary==3.2.3
psycopg2==2.9.10
sqlparse==0.5.2
tzdata==2024.2
